Improved on resizing.
Note that the current implementation assumes that $clock is packed.
This has to be checked (and extended) for grid and place as well.

Also note that the geometry property is set earlier than the
cget (-height/-width) values. The latter cannot be used to check the
current size of the widget when resizing